Minor improvements.
