+++ Assets/GothicVR-Lab/Scripts/LabBootstrapper.cs
@@ -1,9 +1,11 @@
 using System;
 using System.IO;
 using GVR.Bootstrap;
+using GVR.Caches;
 using GVR.Manager.Settings;
 using GVR.Phoenix.Interface;
 using GVR.Phoenix.Interface.Vm;
+using PxCs.Interface;
 using UnityEngine;
 
 namespace GVR.GothicVR_Lab.Scripts
@@ -19,6 +21,25 @@ namespace GVR.GothicVR_Lab.Scripts
 
             var gothicVmPath = Path.GetFullPath($"{SettingsManager.GameSettings.GothicIPath}/_work/DATA/scripts/_compiled/GOTHIC.DAT");
             GameData.VmGothicPtr = VmGothicExternals.LoadVm(gothicVmPath);
+        }
+
+
+        private void OnDestroy()
+        {
+            if (GameData.VfsPtr != IntPtr.Zero)
+            {
+                PxVfs.pxVfsDestroy(GameData.VfsPtr);
+                GameData.VfsPtr = IntPtr.Zero;
+            }
+
+            if (GameData.VmGothicPtr != IntPtr.Zero)
+            {
+                PxVm.pxVmDestroy(GameData.VmGothicPtr);
+                GameData.VmGothicPtr = IntPtr.Zero;
+            }
+
+            GameData.FreePoints.Clear();
+            LookupCache.NpcCache.Clear();
         }
     }
 }
